// webapp/types/backend.ts
// Copyright 2023 mik
// 
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
// 
//     http://www.apache.org/licenses/LICENSE-2.0
// 
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

export enum BackendStatus {
    INIT = 'init',
    WAIT = 'wait',
    STARTING = 'starting',
    STARTED = 'started',
    STOPPING = 'stopping',
    STOPPED = 'stopped',
    ERROR = 'error',
}

export type BackendPayload = {
    status: BackendStatus;
    message?: string;
};

export type BackendContext = {
    server: {
        status: BackendStatus;
        message?: string;
        name?: string;
        stout: string[];
        sterr: string[];
    };
};